BTL_ATBMTT/FlaskProject4/requirements.txt
Flask==2.3.3
flask-socketio==5.3.6
pycryptodome==3.20.0